Move api actions to api/actions folder
Refs #3.
