+++ app/api/actions/units.spec.js
@@ -1,4 +1,4 @@
-import types from './actionTypes';
+import types from '../actionTypes';
 import { createApiTest } from './testUtils';
 import { fetchUnits } from './units';
 import { buildAPIUrl } from './utils';
